OpenAPI YAML definition: Switched from Gist to local definition
The Open API definition found in this Gist
https://gist.github.com/kinlane/6717271a1aaa8e8dec607ff8e7564504
was not parseable, as it defined a single-entry array instead of
an object.

Replaced the remote Gist with the local version of the corrected
OpenAPI definition using Code Prettify library as an alternative.
Now the definition is both parseable and up-to-date with the local
file.